rename color
